requirements.txt
Babel==2.3.4
beautifulsoup4==4.6.0
certifi==2018.1.18
chardet==3.0.4
decorator==4.0.10
Django==2.0.2
django-haystack==2.6.1
docutils==0.12
ebaysdk==2.1.5
feedparser==5.2.1
gevent==1.2.2
greenlet==0.4.10
html2text==2016.9.19
idna==2.6
jieba==0.39
Jinja2==2.8
lxml==4.1.1
Mako==1.0.4
Markdown==2.6.11
MarkupSafe==0.23
mock==2.0.0
num2words==0.5.4
ofxparse==0.16
olefile==0.45.1
passlib==1.6.5
pbr==3.1.1
Pillow==4.0.0
psutil==5.4.3
psycopg2==2.7.3.2
py2exe==0.9.2.2
pydot==1.2.3
pyparsing==2.1.10
PyPDF2==1.26.0
pyserial==3.1.1
python-dateutil==2.5.3
python-stdnum==1.8.1
pytz==2016.7
pyusb==1.0.0
PyYAML==3.12
qrcode==5.3
reportlab==3.3.0
requests==2.11.1
six==1.11.0
suds-jurko==0.6
urllib3==1.22
vatnumber==1.2
vobject==0.9.3
Werkzeug==0.11.15
Whoosh==2.7.4
xlrd==1.0.0
XlsxWriter==0.9.3
xlwt==1.3.0
